clean scarb
