+++ plugins/scarb/src/utils/workspace.ts
